requirements.txt
django
pylint
pylint-django
autopep8
django-mdeditor
django-pure-pagination